+++ infra/conf/transport_internet.go
@@ -24,13 +24,8 @@ import (
 	"github.com/xtls/xray-core/transport/internet"
 	"github.com/xtls/xray-core/transport/internet/finalmask/fragment"
 	"github.com/xtls/xray-core/transport/internet/finalmask/header/custom"
-	"github.com/xtls/xray-core/transport/internet/finalmask/header/dns"
-	"github.com/xtls/xray-core/transport/internet/finalmask/header/dtls"
-	"github.com/xtls/xray-core/transport/internet/finalmask/header/srtp"
-	"github.com/xtls/xray-core/transport/internet/finalmask/header/utp"
-	"github.com/xtls/xray-core/transport/internet/finalmask/header/wechat"
-	"github.com/xtls/xray-core/transport/internet/finalmask/header/wireguard"
 	"github.com/xtls/xray-core/transport/internet/finalmask/mkcp/aes128gcm"
+	"github.com/xtls/xray-core/transport/internet/finalmask/mkcp/header"
 	"github.com/xtls/xray-core/transport/internet/finalmask/mkcp/original"
 	"github.com/xtls/xray-core/transport/internet/finalmask/noise"
 	"github.com/xtls/xray-core/transport/internet/finalmask/realm"
@@ -1243,21 +1238,16 @@ var (
 	}, "type", "settings")
 
 	udpmaskLoader = NewJSONConfigLoader(ConfigCreatorCache{
-		"header-custom":    func() interface{} { return new(HeaderCustomUDP) },
-		"header-dns":       func() interface{} { return new(Dns) },
-		"header-dtls":      func() interface{} { return new(Dtls) },
-		"header-srtp":      func() interface{} { return new(Srtp) },
-		"header-utp":       func() interface{} { return new(Utp) },
-		"header-wechat":    func() interface{} { return new(Wechat) },
-		"header-wireguard": func() interface{} { return new(Wireguard) },
-		"mkcp-original":    func() interface{} { return new(Original) },
-		"mkcp-aes128gcm":   func() interface{} { return new(Aes128Gcm) },
-		"noise":            func() interface{} { return new(NoiseMask) },
-		"salamander":       func() interface{} { return new(Salamander) },
-		"sudoku":           func() interface{} { return new(Sudoku) },
-		"xdns":             func() interface{} { return new(Xdns) },
-		"xicmp":            func() interface{} { return new(Xicmp) },
-		"realm":            func() interface{} { return new(Realm) },
+		"header-custom":  func() interface{} { return new(HeaderCustomUDP) },
+		"mkcp-header":    func() interface{} { return new(Header) },
+		"mkcp-original":  func() interface{} { return new(Original) },
+		"mkcp-aes128gcm": func() interface{} { return new(Aes128Gcm) },
+		"noise":          func() interface{} { return new(NoiseMask) },
+		"salamander":     func() interface{} { return new(Salamander) },
+		"sudoku":         func() interface{} { return new(Sudoku) },
+		"xdns":           func() interface{} { return new(Xdns) },
+		"xicmp":          func() interface{} { return new(Xicmp) },
+		"realm":          func() interface{} { return new(Realm) },
 	}, "type", "settings")
 )
 
@@ -1750,49 +1740,23 @@ func (c *HeaderCustomUDP) Build() (proto.Message, error) {
 	}
 }
 
-type Dns struct {
+type Header struct {
+	ID     int    `json:"id"`
 	Domain string `json:"domain"`
 }
 
-func (c *Dns) Build() (proto.Message, error) {
-	config := &dns.Config{}
-	config.Domain = "www.baidu.com"
-
-	if len(c.Domain) > 0 {
-		config.Domain = c.Domain
+func (c *Header) Build() (proto.Message, error) {
+	if c.ID < 0 || c.ID > 5 {
+		return nil, errors.New("invalid id")
 	}
-
+	config := &header.Config{
+		ID:     int32(c.ID),
+		Domain: c.Domain,
+	}
+	if len(c.Domain) == 0 {
+		config.Domain = "www.baidu.com"
+	}
 	return config, nil
-}
-
-type Dtls struct{}
-
-func (c *Dtls) Build() (proto.Message, error) {
-	return &dtls.Config{}, nil
-}
-
-type Srtp struct{}
-
-func (c *Srtp) Build() (proto.Message, error) {
-	return &srtp.Config{}, nil
-}
-
-type Utp struct{}
-
-func (c *Utp) Build() (proto.Message, error) {
-	return &utp.Config{}, nil
-}
-
-type Wechat struct{}
-
-func (c *Wechat) Build() (proto.Message, error) {
-	return &wechat.Config{}, nil
-}
-
-type Wireguard struct{}
-
-func (c *Wireguard) Build() (proto.Message, error) {
-	return &wireguard.Config{}, nil
 }
 
 type Original struct{}
