--- transport/internet/finalmask/header/dns/config.go
@@ -1,19 +1,0 @@
-package dns
-
-import (
-	"net"
-)
-
-func (c *Config) UDP() {
-}
-
-func (c *Config) WrapPacketConnClient(raw net.PacketConn, level int, levelCount int) (net.PacketConn, error) {
-	return NewConnClient(c, raw)
-}
-
-func (c *Config) WrapPacketConnServer(raw net.PacketConn, level int, levelCount int) (net.PacketConn, error) {
-	return NewConnServer(c, raw)
-}
-
-func (c *Config) HeaderConn() {
-}
