--- transport/internet/finalmask/header/dns/config.proto
@@ -1,11 +1,0 @@
-syntax = "proto3";
-
-package xray.transport.internet.finalmask.header.dns;
-option csharp_namespace = "Xray.Transport.Internet.Finalmask.Header.Dns";
-option go_package = "github.com/xtls/xray-core/transport/internet/finalmask/header/dns";
-option java_package = "com.xray.transport.internet.finalmask.header.dns";
-option java_multiple_files = true;
-
-message Config {
-    string domain = 1;    
-}
