--- transport/internet/finalmask/header/utp/config.proto
@@ -1,9 +1,0 @@
-syntax = "proto3";
-
-package xray.transport.internet.finalmask.header.utp;
-option csharp_namespace = "Xray.Transport.Internet.Finalmask.Header.Utp";
-option go_package = "github.com/xtls/xray-core/transport/internet/finalmask/header/utp";
-option java_package = "com.xray.transport.internet.finalmask.header.utp";
-option java_multiple_files = true;
-
-message Config {}
