+++ transport/internet/finalmask/mkcp/header/config.go
@@ -1,11 +1,12 @@
-package srtp
+package header
 
 import (
 	"net"
 )
 
-func (c *Config) UDP() {
-}
+func (c *Config) UDP() {}
+
+func (c *Config) HeaderConn() {}
 
 func (c *Config) WrapPacketConnClient(raw net.PacketConn, level int, levelCount int) (net.PacketConn, error) {
 	return NewConnClient(c, raw)
@@ -13,7 +14,4 @@ func (c *Config) WrapPacketConnClient(raw net.PacketConn, level int, levelCount 
 
 func (c *Config) WrapPacketConnServer(raw net.PacketConn, level int, levelCount int) (net.PacketConn, error) {
 	return NewConnServer(c, raw)
-}
-
-func (c *Config) HeaderConn() {
 }
