+++ transport/internet/finalmask/mkcp/header/config.proto
@@ -1,0 +1,12 @@
+syntax = "proto3";
+
+package xray.transport.internet.finalmask.mkcp.header;
+option csharp_namespace = "Xray.Transport.Internet.Finalmask.Mkcp.Header";
+option go_package = "github.com/xtls/xray-core/transport/internet/finalmask/mkcp/header";
+option java_package = "com.xray.transport.internet.finalmask.mkcp.header";
+option java_multiple_files = true;
+
+message Config {
+    int32 ID = 1;
+    string domain = 2;
+}
