+++ transport/internet/finalmask/mkcp/header/header.go
@@ -1,0 +1,17 @@
+package header
+
+type Header interface {
+	Size() int
+	Serialize(b []byte)
+}
+
+type HeaderID int
+
+const (
+	DNS HeaderID = iota
+	DTLS
+	SRTP
+	UTP
+	WECHAT
+	WIREGUARD
+)
